Move Correlation stuff to OAuthHandler
