+++ src/Microsoft.AspNet.Authentication.OAuth/Constants.cs
@@ -2,11 +2,11 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 
-namespace Microsoft.AspNet.Authentication
+namespace Microsoft.AspNet.Authentication.OAuth
 {
     internal static class Constants
     {
-        public static string SecurityAuthenticate = "security.Authenticate";
+        internal const string SecurityAuthenticate = "security.Authenticate";
         internal const string CorrelationPrefix = ".AspNet.Correlation.";
     }
 }
